refactor: extract stack-carry simulation into analysis submodule
